add new styling
