+++ static/src/stylesheets/head.identity.scss
@@ -27,3 +27,5 @@
 @import 'module/identity/ad-prefs';
 @import 'module/identity/consent';
 @import 'module/identity/follow';
+@import 'module/identity/upsell/_base';
+@import 'module/identity/upsell/account-creation';
