+++ static/src/stylesheets/module/identity/_forms.scss
@@ -132,6 +132,34 @@ Wrapper
     }
 }
 
+/*
+Interim guui input
+*/
+.identity-forms-input {
+    @include fs-textSans(5);
+    border: 1px solid #dcdcdc;
+    display: block;
+    height: $gs-baseline * 3.5;
+    line-height: $gs-baseline * 1.67;
+    padding: 0 $gs-gutter / 2;
+    width: 100%;
+    transition: all .2s ease-in-out;
+
+    &:hover {
+        box-shadow: 0 0 0 3px #ededed;
+    }
+
+    &:focus {
+      outline: transparent;
+      box-shadow: 0 0 0 3px #ffe500;
+    }
+
+    &[disabled] {
+        background: #EDEDED;
+        pointer-events: none;
+    }
+}
+
 
 
 .identity-forms-wrapper__info {
