+++ static/src/stylesheets/module/identity/upsell/_account-creation.scss
@@ -1,0 +1,15 @@
+.identity-upsell-account-creation-bullets {
+    @include fs-headlineGarnett(0);
+    list-style: none;
+    li {
+        position: relative;
+        margin-bottom: $gs-baseline;
+        &:before {
+            position: absolute;
+            content: '•';
+            color: $green-dark;
+            width: 1em;
+            margin-left: -$gs-gutter;
+        }
+    }
+}
