+++ static/src/stylesheets/module/identity/upsell/_base.scss
@@ -1,0 +1,10 @@
+.identity-upsell-title {
+    @include fs-headlineGarnett(2);
+    font-weight: 600;
+    .identity-upsell-title__title {
+        font-size: 1em;
+    }
+    .identity-upsell-title__subtitle {
+        font-weight: 100;
+    }
+}
